move metrics to metrics file
